adds exisiting functionality from repository to StartupManager and breaks run() into methods for firstRun and startup
