+++ PhotoPoints/StartupManager.swift
@@ -6,14 +6,46 @@
 //  Copyright © 2020 Cascadia College. All rights reserved.
 //
 
-import Foundation
+import CoreData
+import UIKit
 
 class StartupManager {
     
+    // get repository instance for methods requiring data retreival etc
+    static let repository = Repository.instance
+    
+    // get managed context so we can save to core data persistent container
+    static let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
+    
+    static let isFirstRun = true
+    
     static func run(){
-        storeAppConfigData()
+        if isFirstRun {
+            firstRun()
+        }
+        startup()
+    }
+    
+    static func firstRun(){
         createImageAndPhotoDirectories()
-        Repository.instance.loadInitData()
+        loadInitData()
+    }
+    
+    static func startup(){
+        storeAppConfigData()
+    }
+    
+    static func storeAppConfigData(){
+        let defaults = UserDefaults.standard
+        defaults.set(Date(), forKey: "LastUpdated")
+        
+        // somewhere else in app we will have file that contains global constants - ClearOnStartup, DebugModeOn etc
+        defaults.set(true, forKey: "ClearOnStartup")
+        // future keys to be added:
+        
+        // only upload on wifi
+        // API database version
+        
     }
     
     // add image and photo subdirectories to .documents
@@ -31,19 +63,49 @@ class StartupManager {
         }
     }
     
-    static func storeAppConfigData(){
-        let defaults = UserDefaults.standard
-        defaults.set(Date(), forKey: "LastUpdated")
+    static func loadInitData(){
         
-        // future keys to be added:
+        // clear item data from Core Data
+        clearData(entityNames: ["Coordinate", "Detail", "Image", "Item"])
         
-        // only upload on wifi
-        // API database version
+        // build mock database
+        MockDatabase.build()
         
+        // write to core data
+        do {
+            try context.save()
+        } catch let error as NSError {
+            print(error.localizedDescription)
+        }
+        
+        // load images to filesystem
+        loadInitImages()
     }
     
+    // clear all records of an entity from Core Data
+    static func clearData(entityNames: [String]) {
+        for entity in entityNames {
+            let request = NSFetchRequest<NSFetchRequestResult>(entityName: entity)
+            let deleteReq = NSBatchDeleteRequest(fetchRequest: request)
+            do {
+                try context.execute(deleteReq)
+            } catch {
+                print(error)
+            }
+        }
+    }
     
-    
+    // load initial images from xcassets to filesystem to allow us to test getting images
+    // from filesystem, as this is what we'll do for images retrieved from API.
+    static func loadInitImages() {
+        if let items = repository.getItems() {
+            for item in items {
+                if let image = repository.getImageFromXcAssets(item: item), let fileName = repository.getImageNameFromXcAssets(item: item) {
+                    ImageManager.storeImage(image: image, with: fileName, to: .images)
+                }
+            }
+        }
+    }
     
     
 }
